Add support for TLS/HTTPS on both the client and the server
At the moment the server side doesn't support SSL for the web end.
The tty_sender supports TLS for the server connection, which is
specified via a command line argument.

The websockets connection on the frontend side is using an address based
on the window.location.protocol (ws:// or wss://)