Add a way to notify the sender when new receivers connect
Also try to "refresh" the terminal, when new receivers connect, so the
content is fully rendered on the receiver side.
At the moment, the "refresh" is faked with making the window smaller and
then back bigger, after a short time (there doesn't seem to be a way to
tell the app to re-draw itself. Or is there?).